feedback/serializers.py
from rest_framework import serializers

from feedback.models import Feedback
from users.models import Role


class FeedbackSerializer(serializers.ModelSerializer):
    class Meta:
        model = Feedback
        fields = '__all__'

class RoleSerializer(serializers.ModelSerializer):
    class Meta:
        model = Role
        fields = '__all__'