requirements.txt
django==5.1
pillow==11.0.0
psycopg2==2.9.9
djangorestframework==3.15.2
django-rest-swagger==2.2.0